pyproject.toml
[project]
name = "llm-analysis"
version = "0.1.0"
description = "LLM-powered news analysis tool for market sentiment and economic trends"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "click>=8.1.0",
]
